Disable custom build of yaml-cpp version 0.6.3
Version 0.6.2 provided by Ubuntu Focal 20.04 seems to work fine at first glance, and the
way version 0.6.3 is integrated here, in a CMake module, downloads and builds the library
repeatedly for each package that depends on isaac_ros_gxf, and still installs it to
the common install-space instead of keeping it internal to the current package.

Some unrelated packages fail to build with this version with the error

    /usr/bin/ld: cannot find -lnvToolsExt
    collect2: error: ld returned 1 exit status

for library libnvToolsExt.so provided at a non-default location
/usr/local/cuda-11.8/targets/x86_64-linux/lib/libnvToolsExt.so, but apparently the
installed CMake configuration does not contain the proper link directory for this
dependency.